Refactor code smells: reduce parameters and split large class
- Refactored `IngamePreviewRenderer::Render` to use `IngamePreviewOptions` struct, reducing parameter list from 15 to 3.
- Refactored `LiveSocket` `receiveNode` and `sendNode` to use `NodeRequest` and `NodeData` structs, improving signature readability.
- Refactored `IOMapOTBM` by extracting OTBM-specific `Item` serialization logic to new file `source/game/item_otbm.cpp`, reducing `iomap_otbm.cpp` size and improving cohesion.
- Updated `CMakeLists.txt` to include the new file.
