+++ source/live/live_socket.h
@@ -38,6 +38,18 @@ struct LiveCursor {
 	Position pos;
 };
 
+struct NodeRequest {
+	int32_t ndx;
+	int32_t ndy;
+	bool underground;
+};
+
+struct NodeData {
+	int32_t ndx;
+	int32_t ndy;
+	uint32_t floorMask;
+};
+
 class LiveSocket {
 public:
 	LiveSocket();
@@ -69,8 +81,8 @@ public:
 
 protected:
 	// receive / send methods
-	void receiveNode(NetworkMessage& message, Editor& editor, Action* action, int32_t ndx, int32_t ndy, bool underground);
-	void sendNode(uint32_t clientId, MapNode* node, int32_t ndx, int32_t ndy, uint32_t floorMask);
+	void receiveNode(NetworkMessage& message, Editor& editor, Action* action, const NodeRequest& request);
+	void sendNode(uint32_t clientId, MapNode* node, const NodeData& data);
 	void receiveFloor(NetworkMessage& message, Editor& editor, Action* action, int32_t ndx, int32_t ndy, int32_t z, MapNode* node, Floor* floor);
 	void sendFloor(NetworkMessage& message, Floor* floor);
